Move determining review requests and notifies to new notifier class
